DocIntel/requirements.txt
pymupdf
pillow
pytesseract
transformers
python-docx
torchvision
camelot-py
google
openai
opencv-python
google-cloud-vision
ghostscript
easyocr
paddleocr
fpdf2
tabula-py
pdf2image
jpype1
ocrmypdf
grpcio-status==1.48.0
protobuf==3.20.3
kraken
python-bidi
paddlepaddle
